= actor extract StringFormatter from Logging.scala for reuse.
