+++ akka-actor/src/main/scala/akka/event/Logging.scala
@@ -11,8 +11,7 @@ import akka.actor._
 import akka.annotation.{ DoNotInherit, InternalApi }
 import akka.dispatch.RequiresMessageQueue
 import akka.event.Logging._
-import akka.util.ReentrantGuard
-import akka.util.Helpers
+import akka.util.{ Helpers, ReentrantGuard, StringFormatter }
 import akka.{ AkkaException, ConfigurationException }
 
 import scala.annotation.implicitNotFound
@@ -1285,33 +1284,9 @@ trait LoggingAdapter {
    * If `arg` is an `Array` it will be expanded into replacement arguments, which is useful when
    * there are more than four arguments.
    */
-  private def format1(t: String, arg: Any): String = arg match {
-    case a: Array[_] if !a.getClass.getComponentType.isPrimitive ⇒ format(t, a: _*)
-    case a: Array[_] ⇒ format(t, (a map (_.asInstanceOf[AnyRef]): _*))
-    case x ⇒ format(t, x)
-  }
+  private def format1(t: String, arg: Any): String = StringFormatter.formatArray(t, arg)
 
-  def format(t: String, arg: Any*): String = {
-    val sb = new java.lang.StringBuilder(64)
-    var p = 0
-    var startIndex = 0
-    while (p < arg.length) {
-      val index = t.indexOf("{}", startIndex)
-      if (index == -1) {
-        sb.append(t.substring(startIndex, t.length))
-          .append(" WARNING arguments left: ")
-          .append(arg.length - p)
-        p = arg.length
-        startIndex = t.length
-      } else {
-        sb.append(t.substring(startIndex, index))
-          .append(arg(p))
-        startIndex = index + 2
-        p += 1
-      }
-    }
-    sb.append(t.substring(startIndex, t.length)).toString
-  }
+  def format(t: String, arg: Any*): String = StringFormatter.format(t, arg: _*)
 }
 
 /**
